refactor(steps): make all submodules to be private
